Fix warnings for printf-style output warnings
IpfixPrinter and msg() calls could result in warnings because
of missing explicit data conversions. Fixed with this commit

+++ src/common/msg.cc
@@ -115,7 +115,7 @@ extern "C" {
 				threadid = iter->second;
 			}
 
-			printf("%02d:%02d:%02d.%03ld[%d] %6s", tform->tm_hour, tform->tm_min, tform->tm_sec, tv.tv_usec/1000, threadid, MSG_TAB[level]);
+			printf("%02d:%02d:%02d.%03ld[%d] %6s", tform->tm_hour, tform->tm_min, tform->tm_sec, (long)(tv.tv_usec/1000), threadid, MSG_TAB[level]);
 #else
 			//printf("%02d:%02d:%02d.%03ld %6s", tform->tm_hour, tform->tm_min, tform->tm_sec, tv.tv_usec/1000, MSG_TAB[level]);
 			// Gerhard: message level is more important than Milliseconds (at least to me)
@@ -129,7 +129,7 @@ extern "C" {
 
 			if (logtext != NULL) {
 #if defined(DEBUG)
-				snprintf(logtext, EXCEPTION_MAXLEN, "%02d:%02d:%02d.%03ld[%d] %6s", tform->tm_hour, tform->tm_min, tform->tm_sec, tv.tv_usec/1000, threadid, MSG_TAB[level]);
+				snprintf(logtext, EXCEPTION_MAXLEN, "%02d:%02d:%02d.%03ld[%d] %6s", tform->tm_hour, tform->tm_min, tform->tm_sec, (long)(tv.tv_usec/1000), threadid, MSG_TAB[level]);
 #else
 				snprintf(logtext, EXCEPTION_MAXLEN, "%02d:%02d:%02d.%03ld %6s", tform->tm_hour, tform->tm_min, tform->tm_sec, tv.tv_usec/1000, MSG_TAB[level]);
 #endif
